dom/xhr/tests/test_bug1070763.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1405571
-->
<head>
  <title>XMLHttpRequest send data and headers</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank"
   href="https://bugzilla.mozilla.org/show_bug.cgi?id=1405571">Mozilla Bug 1405571</a>
<p id="display">
</p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script class="testbody" type="application/javascript">
SimpleTest.waitForExplicitFinish();

const url = "http://example.com/tests/dom/xhr/tests/file_XHRResponseURL.text";

function runTest(testName, testFn) {
  return new Promise(resolve => {
    const xhr = new XMLHttpRequest();
    xhr.onloadend = () => {
      xhr.onloadend = null;
      xhr.onreadystatechange = () => {
        if (xhr.readyState === 1) {
          testFn(xhr);
        } else if (xhr.readyState === 4) {
          ok(true, testName);
          resolve();
        }
      };
      xhr.open("GET", url);
      xhr.send(null);
    };
    xhr.open("GET", url);
    xhr.send(null);
  });
}

async function runTests() {
  await runTest("Abort #1", xhr => { xhr.abort() });
  await runTest("Abort #2", xhr => { setTimeout(() => xhr.abort(), 0) });
  await runTest("Timeout",  xhr => { xhr.timeout = 1 });
  SimpleTest.finish();
}

runTests();

</script>
</pre>
</body>
</html>
